multiple maintainers via event recieve hook
This is the wrong way to do it, as the state of a trusted maintainers
repo state may not be the same as a maintainer they list.

alice lists bob and carole
bob lists noone

in this scenario alice's repo cant be a symblink to bobs or vice cera
but both should be always up-to-date.

a post-receive hook to update other git repos that list the maintainer
with the new state is simplier and will always be up-to-date.