feat: practively sync newly added repos
when an existing repo adds a new ngit-relay instance, all of the
existing branches arn't synced.

whilst its probably a good idea to add an `ngit sync` feature to
the ngit cli client, this is also a really helpful feature to get
it in sync without doing anything.

it also paves the way for syncing on a timer job and enabling
ngit-relay archives to function.